feat: basic rendered implementation
